AI_Grievance_Analyzer/model/train_model.py
import pandas as pd
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.naive_bayes import MultinomialNB
from sklearn.model_selection import train_test_split
from sklearn.metrics import classification_report
import joblib

# Load dataset
data = pd.read_csv("dataset/grievances1.csv")

# Strip any leading/trailing spaces from column names
data.columns = data.columns.str.strip()

# Make sure the text and department columns exist
print("Columns in dataset:", data.columns)

X = data['Complaint_Text']
y = data['Category']

# Train-test split
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42, stratify=y)

# Vectorize text
vectorizer = TfidfVectorizer(stop_words='english', ngram_range=(1,2))
X_train_vec = vectorizer.fit_transform(X_train)
X_test_vec = vectorizer.transform(X_test)

# Train model
model = MultinomialNB()
model.fit(X_train_vec, y_train)

# Evaluate
y_pred = model.predict(X_test_vec)
print("Classification Report:\n", classification_report(y_test, y_pred))

# Save model and vectorizer
joblib.dump(vectorizer, "vectorizer.pkl")
joblib.dump(model, "grievance_model.pkl")
print("✅ Model and vectorizer saved!")
