AI_Grievance_Analyzer/requirements.txt
openai-whisper
torch
torchvision
torchaudio
easyocr
opencv-python-headless
flask
pandas
scikit-learn
joblib
Flask
requests
python-dotenv
redis
